Only grab mouse, if it has focus.
For example if the window is created, but ois is not initialized and
the user switches to another app, then the whole system gets locked.
Also do not set mouse position. If there is a full screen foreground
window, it will not get inside OIS window, so its not always working.